WS013: Merge apiClient.ts to apiClients.ts
