+++ frontend/src/clients/apiClients.ts
@@ -16,7 +16,12 @@ export interface Species {
 }
 
 export interface Location {
+  id: number;
+  latitude: number;
+  longitude: number;
   name: string;
+  description: string;
+  amenities: string[];
 }
 
 export interface User {
@@ -32,6 +37,28 @@ export interface Sighting {
   species: Species;
   photoUrl: string;
   user: User;
+}
+
+export interface Species {
+  id: number;
+  name: string;
+  latinName: string;
+  photoUrl: string;
+  description: string;
+  endangeredStatus: string;
+}
+
+export interface NewSighting {
+  date: Date;
+  locationId: number;
+  speciesId: number;
+  description: string;
+  photoUrl: string;
+  userId: number;
+}
+
+function getAuthorizationHeader(username: string, password: string) {
+  return `Basic ${btoa(`${username}:${password}`)}`;
 }
 
 export async function GetAllSightings(): Promise<Array<Sighting>> {
@@ -65,6 +92,35 @@ export async function createUser(newUser: NewUser) {
       "Content-Type": "application/json",
     },
     body: JSON.stringify(newUser),
+  });
+
+  if (!response.ok) {
+    throw new Error(await response.json());
+  }
+}
+
+export async function fetchSpecies(): Promise<Array<Species>> {
+  const response = await fetch(`https://localhost:5001/species`);
+  return await response.json();
+}
+
+export async function fetchLocations(): Promise<Array<Location>> {
+  const response = await fetch(`https://localhost:5001/locations`);
+  return await response.json();
+}
+
+export async function createSighting(
+  newSighting: NewSighting,
+  username: string,
+  password: string
+) {
+  const response = await fetch(`https://localhost:5001/sightings/create`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: getAuthorizationHeader(username, password),
+    },
+    body: JSON.stringify(newSighting),
   });
 
   if (!response.ok) {
